+++ core/src/tabledata.h
@@ -58,8 +58,8 @@ class TableData : public QObject
 	Q_OBJECT
 public:
 	TableData();
-	double calcAxis(unsigned short val,QList<QPair<QString,double> > metadata);
-	unsigned short backConvertAxis(double val,QList<QPair<QString,double> > metadata);
+	double calcAxis(int val,QList<QPair<QString,double> > metadata);
+	int backConvertAxis(double val,QList<QPair<QString,double> > metadata);
 };
 
 #endif // TABLEDATA_H
